# Migrates nll_loss_backward from TH to Aten (CUDA)