--- a/aten/src/THCUNN/ClassNLLCriterion.cu
+++ b/aten/src/THCUNN/ClassNLLCriterion.cu
@@ -63,28 +63,6 @@ __global__ void ClassNLLCriterion_updateOutput_no_reduce_kernel(
   }
 }
 
-template <typename Dtype>
-__global__ void ClassNLLCriterion_updateGradInput_no_reduce_kernel(
-    int batch_size,
-    THCDeviceTensor<THCIndex_t, 1> target,
-    THCDeviceTensor<Dtype, 1> gradOutput,
-    THCDeviceTensor<Dtype, 2> gradInput,
-    Dtype *weights,
-    int n_classes,
-    int ignore_index) {
-
-  CUDA_KERNEL_LOOP(index, batch_size) {
-    int cur_target = target[index];
-    if (cur_target == ignore_index) {
-      continue;
-    }
-    CUDA_KERNEL_ASSERT(cur_target  >= 0 && cur_target  < n_classes);
-    Dtype weight =
-       weights ? weights[cur_target] : ScalarConvert<int, Dtype>::to(1);
-    gradInput[index][cur_target] = -weight * gradOutput[index];
-  }
-}
-
 template <typename Dtype, typename Acctype>
 __global__ void cunn_ClassNLLCriterion_updateOutput_kernel(Dtype *output,
                                                            Dtype *total_weight,
@@ -137,56 +115,6 @@ __global__ void cunn_ClassNLLCriterion_updateOutput_kernel(Dtype *output,
       }
     }
 
-  }
-}
-
-template <typename Dtype>
-__global__ void cunn_ClassNLLCriterion_updateGradInput_kernel1(
-  Dtype* gradInput,
-  Dtype* gradOutput,
-  Dtype* weights,
-  THCIndex_t* target,
-  Dtype* total_weight,
-  int size_average,
-  int n_classes,
-  int64_t ignore_index)
-{
-  if (*total_weight <= 0) {
-    return;
-  }
-  Dtype norm = size_average ? (ScalarConvert<int, Dtype>::to(1) / *total_weight) : ScalarConvert<int, Dtype>::to(1);
-  int t = (int)*target;
-  if (t != (int) ignore_index) {
-    CUDA_KERNEL_ASSERT(t >= 0 && t < n_classes);
-    gradInput[t] = -(weights ? weights[t] : ScalarConvert<int, Dtype>::to(1)) * norm * gradOutput[0];
-  }
-}
-
-template <typename Dtype>
-__global__ void cunn_ClassNLLCriterion_updateGradInput_kernel(
-  Dtype *gradInput,
-  Dtype *gradOutput,
-  THCIndex_t *target,
-  Dtype *weights,
-  Dtype *total_weight,
-  int size_average,
-  int nframe,
-  int ndim,
-  int n_classes,
-  int64_t ignore_index)
-{
-  if (*total_weight <= 0) {
-    return;
-  }
-  int i, t;
-  Dtype norm = size_average ? (ScalarConvert<int, Dtype>::to(1) / *total_weight) : ScalarConvert<int, Dtype>::to(1);
-
-  for (i = threadIdx.x; i < nframe; i += NTHREADS) {
-    t = (int)target[i];
-    if (t != (int) ignore_index) {
-      CUDA_KERNEL_ASSERT(t >= 0 && t < n_classes);
-      gradInput[i * ndim + t] = -(weights ? weights[t] : ScalarConvert<int, Dtype>::to(1)) * norm * gradOutput[0];
-    }
   }
 }