… aten/src/ATen/cuda/LegacyTHFunctionsCUDA.cpp
… aten/src/THCUNN/generic/ClassNLLCriterion.cu